Move user key generation to G Suite utils
